refactor: moved function to component
applied feedback from Jesse
